attempt to fix type issue with all-integer arrays

+++ gower/gower_dist.py
@@ -40,7 +40,7 @@ def gower_matrix(data_x, data_y=None, weight=None, cat_features=None):
     x_index = range(0,x_n_rows)
     y_index = range(x_n_rows,x_n_rows+y_n_rows)
     
-    Z_num = Z[:,np.logical_not(cat_features)]
+    Z_num = Z[:,np.logical_not(cat_features)].astype(np.float32)
     
     num_cols = Z_num.shape[1]
     num_ranges = np.zeros(num_cols)
